+++ src/tool/bash-completion.sh
@@ -1,6 +1,7 @@
 _cmdliner_generic() {
   local prefix="${COMP_WORDS[COMP_CWORD]}"
-  COMP_WORDS[COMP_CWORD]="+cmdliner_complete:${COMP_WORDS[COMP_CWORD]}"
+  COMP_WORDS[COMP_CWORD]="--__complete=${COMP_WORDS[COMP_CWORD]}"
+  COMP_WORDS=("${COMP_WORDS[@]:0:1}" "--__complete" "${COMP_WORDS[@]:1}")
   local line="${COMP_WORDS[@]}"
   local version type group item item_doc
   {