+++ test/test_completion.ml
@@ -20,7 +20,7 @@ let complete = Testing_cmdliner.snap_completion cmd
 
 let test_groups =
   Test.test "Cmd.group completions" @@ fun () ->
-  complete ["+cmdliner_complete:"] @@  __POS_OF__
+  complete ["--__complete="] @@  __POS_OF__
     "1\n\
      group\n\
      Subcommands\n\
@@ -36,7 +36,7 @@ let test_groups =
      item\n\
      camels\n\
      Operate on camels.\n";
-  complete ["birds"; "+cmdliner_complete:"] @@  __POS_OF__
+  complete ["birds"; "--__complete="] @@  __POS_OF__
     "1\n\
      group\n\
      Subcommands\n\
@@ -50,14 +50,14 @@ let test_groups =
 
 let test_no_options_after_dashsash =
   Test.test "No options after --" @@ fun () ->
-  complete ["birds"; "fly"; "+cmdliner_complete:"] @@ __POS_OF__
+  complete ["birds"; "fly"; "--__complete="] @@ __POS_OF__
     "1\n\
      group\n\
      Options\n\
      item\n\
      --speed\n\
      Movement \u{001B}[04mSPEED\u{001B}[m in m/s\n";
-  complete ["birds"; "fly"; "--"; "+cmdliner_complete:"] @@ __POS_OF__
+  complete ["birds"; "fly"; "--"; "--__complete="] @@ __POS_OF__
     "1\n";
   ()
 
@@ -79,7 +79,7 @@ let test_restart_restricted_tool =
     ()
   in
   let complete = Testing_cmdliner.snap_completion cmd in
-  complete ["+cmdliner_complete:"] @@ __POS_OF__
+  complete ["--__complete="] @@ __POS_OF__
     "1\n\
      group\n\
      Values\n\
@@ -94,7 +94,7 @@ let test_restart_restricted_tool =
      item\n\
      --verbose\n\
      \n";
-  complete ["+cmdliner_complete:g"] @@ __POS_OF__
+  complete ["--__complete=g"] @@ __POS_OF__
     "1\n\
      group\n\
      Values\n\
@@ -104,14 +104,14 @@ let test_restart_restricted_tool =
      group\n\
      Options\n";
   (* Note no reset here: as there is no -- token *)
-  complete ["git"; "+cmdliner_complete:"] @@ __POS_OF__
+  complete ["git"; "--__complete="] @@ __POS_OF__
     "1\n\
      group\n\
      Options\n\
      item\n\
      --verbose\n\
      \n";
-  complete ["--"; "git"; "+cmdliner_complete:"] @@ __POS_OF__
+  complete ["--"; "git"; "--__complete="] @@ __POS_OF__
     "1\n\
      restart\n";
   ()
@@ -130,7 +130,7 @@ let test_restart_any_tool =
     ()
   in
   let complete = Testing_cmdliner.snap_completion cmd in
-  complete ["+cmdliner_complete:"] @@ __POS_OF__
+  complete ["--__complete="] @@ __POS_OF__
     "1\n\
      group\n\
      Options\n\
@@ -138,11 +138,11 @@ let test_restart_any_tool =
      --verbose\n\
      \n";
   (* The following two do not restart because -- is missing *)
-  complete ["+cmdliner_complete:gi"] @@ __POS_OF__
+  complete ["--__complete=gi"] @@ __POS_OF__
     "1\n\
      group\n\
      Options\n";
-  complete ["git"; "+cmdliner_complete:"] @@ __POS_OF__
+  complete ["git"; "--__complete="] @@ __POS_OF__
     "1\n\
      group\n\
      Options\n\
@@ -150,10 +150,10 @@ let test_restart_any_tool =
      --verbose\n\
      \n";
   (* These must restart *)
-  complete ["--"; "+cmdliner_complete:gi"] @@ __POS_OF__
+  complete ["--"; "--__complete=gi"] @@ __POS_OF__
     "1\n\
      restart\n";
-  complete ["--"; "git"; "+cmdliner_complete:"] @@ __POS_OF__
+  complete ["--"; "git"; "--__complete="] @@ __POS_OF__
     "1\n\
      restart\n";
   ()